// pkg/packages/gcp/src/generated/run-jobs/model/googleCloudRunV2EnvVar.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cloud Run Admin API
 * OpenAPI spec version: v2
 */
import type { GoogleCloudRunV2EnvVarSource } from './googleCloudRunV2EnvVarSource';

/**
 * EnvVar represents an environment variable present in a Container.
 */
export interface GoogleCloudRunV2EnvVar {
  /** Required. Name of the environment variable. Must not exceed 32768 characters. */
  name?: string;
  /** Literal value of the environment variable. Defaults to "", and the maximum length is 32768 bytes. Variable references are not supported in Cloud Run. */
  value?: string;
  valueSource?: GoogleCloudRunV2EnvVarSource;
}